Add createEntitySelectors
This method creates a set of frequently needed
selectors based on a passed entity state.

- getIds and getEntities return all ids and entities
stored of the entity state.
- getAll returns all included entities as array
- getFirstSelected returns the first selected entity or null
- isEntitySelected returns whether the entity state has selected entities

We also bump versions to 2.0.0 and set isPrivate to true in dist/package.json.
This commit marks the beginning of another version with some major changes.
For instance, this package will need some dependencies such as 'reselect'.