Add function-based action creators and (optionally) type actions
entity-store so far has offered class-based action creators.
The convetions in redux and ngrx have adopted
function-based action creators, so we should offer this as well.

The old style is kept as well, so full compatibility is ensured.

CompositeEntityActions can now be typed with an EntityTypeMap.
This ensures that users get typing support when filling in entity types
and payloads.